SourceCode and files
The "Youth Challenge Data and Viz.xlsx" file contains all the data along with graphs. Sorted by sheets.

The "Youth Offence Graph" graph was produced using R in Rstudio. 
The RMarkdown file and output HTML file are also attached as
 "Youth Challenge Viz.Rmd" and "Youth-Challenge-Viz.html " respectively.
The data that was in R is also in the Excel file above.